requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-heroku==0.3.1
grep==0.3.2
gunicorn==20.1.0
nano==0.10.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
whitenoise==6.4.0
